fix(glm): GLM prompt-built DSA state closes the KV-only prefix cache
`kv_only_prefix_cache_is_safe()` refused compressed DeepSeek V4 because its
compressor pool and ring are prompt-built and KV blocks do not carry them.
GLM-5.3 has the same shape and was never taught to the predicate: prefill
builds `Glm5NextDsaState` (the DSA indexer rows) per sequence, and a radix
prefix hit reuses KV blocks against an indexer that was never populated for
that prefix. `--enable-prefix-caching` on a `glm5_next` serve is therefore a
silently wrong answer, not a crash.

The predicate becomes a `match` over `model_type` so a third model is one
arm, and `glm5_next` / `glm5_next_text` join it. Behaviour for every other
model is byte-identical. The predicate stays the single preflight SSOT that
`build_prefix_cache` consults; the warning it logs no longer names DeepSeek
V4 and carries `model_type` instead.

The `methods.rs` test module moves to a sibling `kv_completeness_tests.rs`
rather than growing in place — `methods.rs` was 459 LoC against the 500-LoC
cap, and the parser file that would otherwise host an end-to-end assertion
(`config/parsers/glm5_next.rs`) is already 972 LoC. The parser's existing
`assert_eq!(c.model_type, "glm5_next")` plus these tests pin the same chain
without adding a line to either over-cap file.

Tests: cargo test -p atlas-core --lib config::kv_completeness (3 passed)
       cargo test -p spark-server --bins -- prefix_cache_tests (3 passed)

(cherry picked from commit 3cf787bd6a2d556fb6ecc5ba47b996a6087c3dae)
